chore: rename "keywords.js" to "constants.js"
... for clarity
